Add Recoverer
