+++ lib/core/src/lib.rs
@@ -176,8 +176,7 @@ pub mod logger;
 pub mod model;
 pub mod persist;
 pub(crate) mod receive_swap;
-#[allow(dead_code)]
-mod restore;
+pub(crate) mod recover;
 pub mod sdk;
 pub(crate) mod send_swap;
 pub(crate) mod signer;
