+++ lib/core/src/recover/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod model;
+pub(crate) mod recoverer;
